+++ exercise33.cpp
@@ -4,15 +4,15 @@
 #include <cstdlib>
 
 /* Guide to Scientific Computing in C++
- * Chapter 3.  Exercise 3.2 */
+ * Chapter 3.  Exercise 3.3 */
 
 int main (int argc, char* argv[])
 {
     int N = atoi(argv[1]);  // Grid size entered on command line
     assert(N > 1);
     
-    float x_prev = 1, x_next = 1;
-    float h = 1.0/N;   // the step size
+    double x_prev = 1.0, x_next = 1.0;
+    double h = 1.0 / (double) N;   // the step size
     std::cout << h << "\n";
     std::ofstream write_file("xy.dat");
     assert(write_file.is_open());
@@ -22,7 +22,7 @@ int main (int argc, char* argv[])
     {
         x_prev = x_next;
         x_next = x_prev - (h * x_prev);
-        write_file << (float) i * h << "  " << x_next << "\n";
+        write_file << (double)i * h << "  " << x_next << "\n";
     }
     
     write_file.close();
